+++ changelog.md
@@ -3,3 +3,7 @@
 - paco : Mario sprite
 - paco : Project structure (scenes folder, spearate player & scene)
 - paco : create default level scene, example player & platform
+
+## paco/mushroom
+
+- Create mushroom scene, WIP mushroom collision code, TODO put it in mushroom script